Specify convolution type="open" in convolve()
The default setting for the convolution type in the convolve() function is set to circular (R Stats Package version 3.7.0), which leads to the unwanted effect that signals happening towards the end of the time sequence and exceeding it, are continued at the beginning due to being treated as periodic. Specifying the convolution type as "open" solves this problem.