Rescaling all noise components for weights is mixture.
To add different noise components in a mixture setting,
each variance component needs to be rescaled.

Some of the noise components were lacking this rescaling step.
Added this here.

NOTE:
Rescaling is done over the 4D image, not within
each voxel seperately.
In other words, an image of 100 voxels is a Monte Carlo simulation
of 100 iterations.